+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.10
 - WdsTextArea > 최소 높이 60px / 최대 높이 320px 대응될 수 있게끔 개선
+- WdsTextArea > 'label' 파라미터 String에서 Text로 형변환 
 
 
 ### 0.0.9